Move chat route tests into routes folder
